Encoder moved to separate file
