+++ memcrs/src/protocol/binary/connection.rs
@@ -1,5 +1,8 @@
 use crate::protocol::binary::codec::{
-    BinaryRequest, BinaryResponse, MemcacheBinaryCodec, ResponseMessage,
+    BinaryRequest, MemcacheBinaryCodec, 
+};
+use crate::protocol::binary::encoder::{
+     BinaryResponse, MemcacheBinaryEncoder, ResponseMessage,
 };
 use bytes::BytesMut;
 use std::cmp;
@@ -11,7 +14,8 @@ use tokio_util::codec::Decoder;
 
 pub struct MemcacheBinaryConnection {
     stream: TcpStream,
-    codec: MemcacheBinaryCodec,
+    decoder: MemcacheBinaryCodec,
+    encoder: MemcacheBinaryEncoder,
     buffer: BytesMut,
 }
 
@@ -19,7 +23,8 @@ impl MemcacheBinaryConnection {
     pub fn new(socket: TcpStream, item_size_limit: u32) -> Self {
         MemcacheBinaryConnection {
             stream: socket,
-            codec: MemcacheBinaryCodec::new(item_size_limit),
+            decoder: MemcacheBinaryCodec::new(item_size_limit),
+            encoder: MemcacheBinaryEncoder::new(),
             buffer: BytesMut::with_capacity(4096),
         }
     }
@@ -29,7 +34,7 @@ impl MemcacheBinaryConnection {
         loop {
             // Attempt to parse a frame from the buffered data. If enough data
             // has been buffered, the frame is returned.
-            if let Some(frame) = self.codec.decode(&mut self.buffer)? {
+            if let Some(frame) = self.decoder.decode(&mut self.buffer)? {
                 match frame {
                     BinaryRequest::ItemTooLarge(request) => {
                         debug!(
@@ -127,7 +132,7 @@ impl MemcacheBinaryConnection {
     }
 
     pub async fn write(&mut self, msg: &BinaryResponse) -> io::Result<()> {
-        let message = self.codec.encode_message(msg);
+        let message = self.encoder.encode_message(msg);
         self.write_data_to_stream(message).await?;
         Ok(())
     }
