+++ memcrs/src/protocol/binary/encoder/binary_encoder_tests.rs
@@ -18,7 +18,7 @@ mod tests {
     }
 
     fn encode_packet(src: BinaryResponse) -> Result<BytesMut, io::Error> {
-        let mut encoder = MemcacheBinaryCodec::new(1024);
+        let mut encoder = MemcacheBinaryEncoder::new();
         let mut buf = BytesMut::with_capacity(128);
         match encoder.encode(src, &mut buf) {
             Ok(_) => Ok(buf),
