+++ memcrs/src/protocol/binary/mod.rs
@@ -1,3 +1,4 @@
 pub mod binary;
 pub mod codec;
+pub mod encoder;
 pub mod connection;
